requirements.txt
# Greek News Aggregator - Dependencies
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
python-dateutil==2.8.2
jinja2==3.1.3
anthropic>=0.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
lunr>=0.7.0
PyYAML>=6.0
aiohttp>=3.9.0
nh3>=0.2.14
pydantic>=2.0
pydantic-settings>=2.0
